Removed stringstreams.
This is done to get rid of the heap dependency.  This library aims to be
used in embedded environments, where dynamic memory allocation is not
always available.  Therefore I am working on replacing all functions
that depend on dynamic memory with static ones.

At this point we got rid of stringstreams. strings are next.